requirements.txt
django-admin-sortable2>=0.6.16
